Workaround Linux pre-4.8-rc1 bug with hwrng
Somehow the GPG key generation process was still blocking. It turns out
that the initialization of the nonblocking pool was not really making
progress due to a bug in the random subsystem of the kernel.

Add a workaround to complete the initialization. After that, a kernel
thread will add entropy from the hwrng device so no rngd is needed.